# docs/swagger.yaml
basePath: /fab
definitions:
...
  gintool.ApiResponse:
    properties:
      code:
        description: 状态码
        type: integer
      err_msg:
        description: 内部错误详情
        type: string
      message:
        description: 状态短语
        type: string
      result:
        description: 数据结果集
        type: object
    type: object
...
  log.setLogLevelRequest:
    properties:
      loglevel:
        example: info
        type: string
    required:
    - loglevel
    type: object
host: localhost:6922
info:
  contact: {}
  description: 描述:fabapi 接口文档.
  license: {}
  title: fabapi API文档
  version: 1.1.0
paths:
...
  /log/setloglevel:
    post:
      consumes:
      - application/json
      description: 功能：设置日志级别
      parameters:
      - description: 设置日志级别
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/log.setLogLevelRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 设置日志级别
      tags:
      - 设置日志级别
swagger: "2.0"